// packages/api/payments/infrastructure/constants/environment.ts
export enum PaymentsStackEnvironmentKeys {
  STRIPE_SECRET_KEY = 'STRIPE_SECRET_KEY',
  STRIPE_PUBLIHABLE_KEY = 'STRIPE_PUBLISHABLE_KEY',
  CURRENCY_CODE = 'CURRENCY_CODE', // AUD or GBP
  TRANSACTIONS_EVENTS_TABLE = 'TRANSACTIONS_EVENTS_TABLE', // AUD or GBP
  PAYMENTS_WEB_HOST = 'PAYMENTS_WEB_HOST',
  DD_VERSION = 'DD_VERSION',
  DD_ENV = 'DD_ENV',
  DD_API_KEY = 'DD_API_KEY',
  DD_GIT_COMMIT_SHA = 'DD_GIT_COMMIT_SHA',
  DD_GIT_REPOSITORY_URL = 'DD_GIT_REPOSITORY_URL',

  // Environment
  USE_DATADOG_AGENT = 'USE_DATADOG_AGENT',

  SKIP_PAYMENTS_STACK = 'false',
}